Update translation to avoid interpolation warnings
We were getting an inconsistent interpolation warning for
every language except the default one, English.

The Crowdin translation interface only shows the English
`one` and `other` plural forms. However, it does not
show the English translation for the zero plural form,
which we translated as an empty string in this particular
case. Because Crowdin translators cannot see our zero
form translation, an empty string, they usually add a
translation for the zero plural form in their language.

The i18n-task `check-consistent-interpolations` tool
checks if the zero plural form of each language translation
contains the same interpolations variables that the zero
form of the source language. In this case, the tool
compares an empty string with the translation for which
usually will contain the %{count} interpolation variable.